zendesk#980 hide timer in review
